chore: improve ts, all url polyfill (#62)
* Some `ts-ignores` are fixed, some just changed to something less
dangerous:
E.g. unlike @ts-ignore, @ts-expect-error throws an error on ts-lint
check if error is not there anymore. So at some point the types for
`Animated.FlatList` will be fixes, we'll know about it and remove the
comment completely

* Use `react-native-url-polyfill` instead of custom `url-parse` lib
since polyfill uses build in react-native modules to deal with urls
instead of parsing using regexps in `url-parse`. And it is easier to
switch from a polyfill if at some point RN adds the support out of the
box

* Fix the issue when on other user's profile the username of the main
user is displayed