chore: saga watchers (#85)
* Fix android `assembleRelease` build (Release master production builds)
* Change the way we export sagas:

Previously we exported module sagas this way:
```
export function* rootModuleSaga () {
   yield all([
       takeLatest(actionA, sagaA),
       takeLatest(actionB, sagaB),
   ])
}
```

and spawned each root module saga later in `rootSaga.ts`

Problem: if error happens in sagaA, the whole rootModuleSaga is
restarted (restarting both watchers - `takeLatest(actionA, sagaA)` and
`takeLatest(actionB, sagaB)`. That is overhead and in some cases leads
to unpredicted behaviour - e.g. when on auth we stuck if error happens
in `userStateChangeSaga`.

How to solve - export array of watchers and spawn each of them
separately in `rootSaga.ts`, so error in any of saga won't affect other
sagas.